# Hack-Nation/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PRISM inference API
  summary: Research-prototype PCOS evidence profiling. Not a diagnostic device.
  version: 0.1.0
paths:
...
  /api/v1/jobs/speech:
    post:
      tags:
      - jobs
      summary: Create Speech Job
      operationId: create_speech_job_api_v1_jobs_speech_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSubmission'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/speech/{job_id}:
    get:
      tags:
      - jobs
      summary: Get Speech Job
      operationId: get_speech_job_api_v1_jobs_speech__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/speech/transcribe:
    post:
      tags:
      - speech
      summary: Transcribe
      description: 'Transcribe an upload and extract proposed clinical events.


        Every returned event is ``proposed``. Extraction is rule-based and reads

        negation and family-versus-patient attribution, but it is not reliable

        enough to enter the record unreviewed, so confirmation is the UI''s job.'
      operationId: transcribe_api_v1_speech_transcribe_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_transcribe_api_v1_speech_transcribe_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Transcribe Api V1 Speech Transcribe Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    Body_transcribe_api_v1_speech_transcribe_post:
      properties:
        audio:
          type: string
          contentMediaType: application/octet-stream
          title: Audio
        patient_id:
          type: string
          title: Patient Id
          default: demo-patient
        language:
          type: string
          title: Language
          default: en
      type: object
      required:
      - audio
      title: Body_transcribe_api_v1_speech_transcribe_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JobRecord:
      properties:
        job_id:
          type: string
          title: Job Id
        kind:
          type: string
          enum:
          - documents
          - speech
          - ultrasound
          title: Kind
        patient_id:
          type: string
          title: Patient Id
        status:
          type: string
          enum:
          - queued
          - processing
          - completed
          - failed
          - unavailable
          title: Status
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
      additionalProperties: false
      type: object
      required:
      - job_id
      - kind
      - patient_id
      - status
      - created_at
      - updated_at
      title: JobRecord
      description: A job's observable state.
    JobSubmission:
      properties:
        patient_id:
          type: string
          minLength: 1
          title: Patient Id
        source_ids:
          items:
            type: string
          type: array
          title: Source Ids
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - patient_id
      title: JobSubmission
      description: What a client sends to open a job.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...